microservices/notificator-service/requirements.txt
pydantic
pydantic-settings
fastapi
aiogram
python-dotenv
redis
uvloop
celery
flower